feat: 4 tools are included in the first commit.
`binary2HDF5` is the tool to convert the binary files generated from Hercules to HDF5 format.
The binary file type can be `plane` (the displacements at points on the specified plane) or `mesh` (The coordinates of all the points in the mesh and the velocity profile of each element).
`makeAnimation` creates an animation of the displacement or velocity propagation across a plane, which is saved in an HDF5 file (it can be generated with the `binary2HDF5` tool).
`plotStations` is used to plot a series of time history (displacement, velocity, and acceleration) at stations recorded as station files (station.0, station.1, etc. by Hercules).
`plotVelocityProfile` is the tool to plot an arbitrary x-z or y-z plane showing the velocity profile in the model. It can also plot the corresponding velocity profile of a 3D velocity model.