refactor: Reduced duplicate lines and fine-tuned components on the figure
The duplicate lines used for generating the first frame and the following frames are reduced.
Using the fixed-width font, Courier New, to make sure the title is not changing from frame to frame (since the timestamp is in the title).
Setting up the minimum threshold for the displacement and rounding the maximum displacement to make sure the width occupation is the same for velocity and displacement.